remove duplicate composable
